// src/generated/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type HeatMap = {
  __typename?: 'HeatMap';
  data: Array<Scalars['Int']>;
  end: Scalars['Date'];
  start: Scalars['Date'];
};
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  /** The last cursor in the page. */
  endCursor: Maybe<Scalars['String']>;
  /** Is there another page after. */
  hasNextPage: Scalars['Boolean'];
  /** Is there a preceding page. */
  hasPreviousPage: Scalars['Boolean'];
  /** The first cursor in the page. */
  startCursor: Maybe<Scalars['String']>;
};

export type Query = {
  __typename?: 'Query';
  /** Test Connection */
  test: Scalars['Boolean'];
  /** Try to find one user */
  user: User;
};

export type QueryUserArgs = {
  username: Scalars['String'];
};

/** One Target */
export type Target = {
  __typename?: 'Target';
  /** Create Time */
  created_at: Scalars['DateTime'];
  /** Generate HeatMap */
  heatMap: HeatMap;
  /** Target ID */
  id: Scalars['ID'];
  /** Timepieces in the last week */
  lastWeekTimePieces: Array<TimePiece>;
  /** Target name */
  name: Scalars['String'];
  /** Total pomodoro count */
  pomodoroCount: Scalars['Int'];
  /** Consumption time periods */
  timePieces: TimePieceConnection;
  /** Time spent on this target */
  timeSpent: Scalars['Int'];
  /** Today pomodoro count */
  todayPomodoroCount: Scalars['Int'];
  /** Today focus time length */
  todayTimeSpent: Scalars['Int'];
  /** Target type */
  type: Maybe<TargetType>;
};

/** One Target */
export type TargetHeatMapArgs = {
  end: InputMaybe<Scalars['Date']>;
};

/** One Target */
export type TargetTimePiecesArgs = {
  after: InputMaybe<Scalars['String']>;
  first: Scalars['Int'];
};

export type TargetCreateInput = {
  name: Scalars['String'];
  timePieces: InputMaybe<Array<TimePieceCreateInput>>;
  timeSpent: InputMaybe<Scalars['Int']>;
  type: InputMaybe<TargetType>;
};

export enum TargetType {
  Longterm = 'LONGTERM',
  Normal = 'NORMAL',
}

export type TargetUpdateInput = {
  name: InputMaybe<Scalars['String']>;
  timeSpent: InputMaybe<Scalars['Int']>;
  type: InputMaybe<TargetType>;
};

/** One Time Piece */
export type TimePiece = {
  __typename?: 'TimePiece';
  /** Time spend on this piece, count as seconds */
  duration: Scalars['Int'];
  /** TimePiece ID */
  id: Scalars['Int'];
  /** Piece start time */
  start: Scalars['DateTime'];
  /** Target this piece belongs to */
  target: Target;
  /** Timepiece type */
  type: Maybe<TimePieceType>;
};

export type TimePieceConnection = {
  __typename?: 'TimePieceConnection';
  edges: Array<TimePieceEdge>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int'];
};

export type TimePieceCreateInput = {
  duration: Scalars['Int'];
  start: Scalars['DateTime'];
  type: InputMaybe<TimePieceType>;
};

export type TimePieceEdge = {
  __typename?: 'TimePieceEdge';
  cursor: Scalars['String'];
  node: TimePiece;
};

export enum TimePieceType {
  Normal = 'NORMAL',
  Pomodoro = 'POMODORO',
}

export type TimePieceUpdateInput = {
  duration: InputMaybe<Scalars['Int']>;
  start: InputMaybe<Scalars['DateTime']>;
  type: InputMaybe<TimePieceType>;
};

/** User */
export type User = {
  __typename?: 'User';
  /** Create Time */
  created_at: Scalars['DateTime'];
  /** Generate heatMap */
  heatMap: HeatMap;
  /** User ID */
  id: Scalars['ID'];
  /** Timepieces in the last week */
  lastWeekTimePieces: Array<TimePiece>;
  /** Total pomodoro count */
  pomodoroCount: Scalars['Int'];
  /** Find a target by ID */
  target: Target;
  /** Targets */
  targets: Maybe<Array<Target>>;
  /** Consumption time periods */
  timePieces: TimePieceConnection;
  /** Total focus time length */
  timeSpent: Scalars['Int'];
  /** Today pomodoro count */
  todayPomodoroCount: Scalars['Int'];
  /** Today focus time length */
  todayTimeSpent: Scalars['Int'];
  /** Username */
  username: Scalars['String'];
};

/** User */
export type UserHeatMapArgs = {
  end: InputMaybe<Scalars['Date']>;
};

/** User */
export type UserTargetArgs = {
  id: Scalars['ID'];
};

/** User */
export type UserTimePiecesArgs = {
  after: InputMaybe<Scalars['String']>;
  first: Scalars['Int'];
};

export type UserCreateInput = {
  targets: InputMaybe<Array<TargetCreateInput>>;
  username: Scalars['String'];
};

export type UserUpdateInput = {
  username: InputMaybe<Scalars['String']>;
};